+++ randovania/gui/tracker_window.py
@@ -1,21 +1,30 @@
+import collections
 import functools
 import json
+import typing
 from pathlib import Path
+from random import Random
 from typing import Optional, Dict, Set, List, Tuple, Iterator, Union
 
+from PySide2 import QtWidgets
 from PySide2.QtCore import Qt
 from PySide2.QtWidgets import QMainWindow, QTreeWidgetItem, QCheckBox, QLabel, QGridLayout, QWidget, QMessageBox
 
+from randovania.game_description.area_location import AreaLocation
 from randovania.game_description.game_description import GameDescription
 from randovania.game_description.item.item_category import ItemCategory
-from randovania.game_description.node import Node, ResourceNode, TranslatorGateNode
+from randovania.game_description.node import Node, ResourceNode, TranslatorGateNode, TeleporterNode
+from randovania.game_description.resources.item_resource_info import ItemResourceInfo
 from randovania.game_description.resources.pickup_entry import PickupEntry
 from randovania.game_description.resources.resource_info import add_resource_gain_to_current_resources
-from randovania.generator import base_patches_factory, generator
+from randovania.game_description.resources.translator_gate import TranslatorGate
+from randovania.generator import generator
 from randovania.gui.generated.tracker_window_ui import Ui_TrackerWindow
 from randovania.gui.lib.common_qt_lib import set_default_window_icon
 from randovania.gui.lib.custom_spin_box import CustomSpinBox
-from randovania.layout.layout_configuration import LayoutConfiguration
+from randovania.layout import translator_configuration
+from randovania.layout.layout_configuration import LayoutConfiguration, LayoutElevators
+from randovania.layout.translator_configuration import LayoutTranslatorRequirement
 from randovania.resolver.bootstrap import logic_bootstrap
 from randovania.resolver.logic import Logic
 from randovania.resolver.resolver_reach import ResolverReach
@@ -58,6 +67,10 @@ class TrackerWindow(QMainWindow, Ui_TrackerWindow):
     layout_configuration: LayoutConfiguration
     persistence_path: Path
     _initial_state: State
+    _elevator_id_to_combo: Dict[int, QtWidgets.QComboBox]
+    _translator_gate_to_combo: Dict[TranslatorGate, QtWidgets.QComboBox]
+    _starting_nodes: Set[ResourceNode]
+    _undefined_item = ItemResourceInfo(-1, "Undefined", "Undefined", 0, None)
 
     # UI tools
     _asset_id_to_item: Dict[int, QTreeWidgetItem]
@@ -78,13 +91,7 @@ class TrackerWindow(QMainWindow, Ui_TrackerWindow):
         self.layout_configuration = layout_configuration
         self.persistence_path = persistence_path
 
-        player_index = 0  # FIXME
-        try:
-            player_pool = generator.create_player_pool(None, self.layout_configuration, player_index)
-        except base_patches_factory.MissingRng as e:
-            raise InvalidLayoutForTracker("Layout is configured to have random {}, "
-                                          "but that isn't supported by the tracker.".format(e))
-
+        player_pool = generator.create_player_pool(Random(0), self.layout_configuration, 0)
         pool_patches = player_pool.patches
         self.game_description, self._initial_state = logic_bootstrap(layout_configuration,
                                                                      player_pool.game,
@@ -98,7 +105,7 @@ class TrackerWindow(QMainWindow, Ui_TrackerWindow):
         self.hide_collected_resources_check.stateChanged.connect(self.update_locations_tree_for_reachable_nodes)
         self.undo_last_action_button.clicked.connect(self._undo_last_action)
 
-        self.configuration_label.setText("Trick Level: {}; Elevators: Vanilla; Starts with:\n{}".format(
+        self.configuration_label.setText("Trick Level: {}; Starts with:\n{}".format(
             layout_configuration.trick_level_configuration.pretty_description,
             ", ".join(
                 resource.short_name
@@ -108,17 +115,15 @@ class TrackerWindow(QMainWindow, Ui_TrackerWindow):
 
         self.setup_pickups_box(player_pool.pickups)
         self.setup_possible_locations_tree()
-
-        self._starting_nodes = {
-            node
-            for node in self.game_description.world_list.all_nodes
-            if node.is_resource_node and node.resource() in self._initial_state.resources
-        }
+        self.setup_elevators()
+        self.setup_translator_gates()
 
         persistence_path.mkdir(parents=True, exist_ok=True)
         previous_state = _load_previous_state(persistence_path, layout_configuration)
 
         if not self.apply_previous_state(previous_state):
+            self.setup_starting_location(None)
+
             with persistence_path.joinpath("layout_configuration.json").open("w") as layout_file:
                 json.dump(layout_configuration.as_json, layout_file)
             self._add_new_action(self._initial_state.node)
@@ -126,6 +131,10 @@ class TrackerWindow(QMainWindow, Ui_TrackerWindow):
     def apply_previous_state(self, previous_state: Optional[dict]) -> bool:
         if previous_state is None:
             return False
+
+        starting_location = None
+        needs_starting_location = len(self.layout_configuration.starting_location.locations) > 1
+        resource_db = self.game_description.resource_database
 
         try:
             pickup_name_to_pickup = {pickup.name: pickup for pickup in self._collected_pickups.keys()}
@@ -137,8 +146,40 @@ class TrackerWindow(QMainWindow, Ui_TrackerWindow):
                 self.game_description.world_list.all_nodes[index]
                 for index in previous_state["actions"]
             ]
+            if needs_starting_location:
+                starting_location = AreaLocation.from_json(previous_state["starting_location"])
+
+            elevators = {
+                int(elevator_id): AreaLocation.from_json(location) if location is not None else None
+                for elevator_id, location in previous_state["elevators"].items()
+            }
+            translator_gates = {
+                TranslatorGate(int(gate)): (resource_db.get_item(item)
+                                            if item is not None
+                                            else self._undefined_item)
+                for gate, item in previous_state["translator_gates"].items()
+            }
         except KeyError:
             return False
+
+        self.setup_starting_location(starting_location)
+
+        for elevator_id, area_location in elevators.items():
+            combo = self._elevator_id_to_combo[elevator_id]
+            if area_location is None:
+                combo.setCurrentIndex(0)
+                continue
+            for i in range(combo.count()):
+                if area_location == combo.itemData(i):
+                    combo.setCurrentIndex(i)
+                    break
+
+        for gate, item in translator_gates.items():
+            combo = self._translator_gate_to_combo[gate]
+            for i in range(combo.count()):
+                if item == combo.itemData(i):
+                    combo.setCurrentIndex(i)
+                    break
 
         self.bulk_change_quantity(quantity_to_change)
         self._add_new_actions(previous_actions)
@@ -153,6 +194,11 @@ class TrackerWindow(QMainWindow, Ui_TrackerWindow):
         while len(self._actions) > 1:
             self._actions.pop()
             self.actions_list.takeItem(len(self._actions))
+
+        for elevator in self._elevator_id_to_combo.values():
+            elevator.setCurrentIndex(0)
+        for elevator in self._translator_gate_to_combo.values():
+            elevator.setCurrentIndex(0)
 
         self._refresh_for_new_action()
 
@@ -227,7 +273,7 @@ class TrackerWindow(QMainWindow, Ui_TrackerWindow):
                     node_item = self._node_to_item[node]
                     node_item.setHidden(not is_visible)
                     if node.is_resource_node:
-                        resource_node: ResourceNode = node
+                        resource_node = typing.cast(ResourceNode, node)
                         node_item.setDisabled(not resource_node.can_collect(state.patches, state.resources))
                         node_item.setCheckState(0, Qt.Checked if is_collected else Qt.Unchecked)
 
@@ -235,6 +281,10 @@ class TrackerWindow(QMainWindow, Ui_TrackerWindow):
                 self._asset_id_to_item[area.area_asset_id].setHidden(not area_is_visible)
 
         # Persist the current state
+        self.persist_current_state()
+
+    def persist_current_state(self):
+        world_list = self.game_description.world_list
         with self.persistence_path.joinpath("state.json").open("w") as state_file:
             json.dump(
                 {
@@ -245,7 +295,16 @@ class TrackerWindow(QMainWindow, Ui_TrackerWindow):
                     "collected_pickups": {
                         pickup.name: quantity
                         for pickup, quantity in self._collected_pickups.items()
-                    }
+                    },
+                    "elevators": {
+                        str(elevator_id): combo.currentData().as_json if combo.currentIndex() > 0 else None
+                        for elevator_id, combo in self._elevator_id_to_combo.items()
+                    },
+                    "translator_gates": {
+                        str(gate.index): combo.currentData().index if combo.currentIndex() > 0 else None
+                        for gate, combo in self._translator_gate_to_combo.items()
+                    },
+                    "starting_location": world_list.node_to_area_location(self._initial_state.node).as_json,
                 },
                 state_file
             )
@@ -273,14 +332,123 @@ class TrackerWindow(QMainWindow, Ui_TrackerWindow):
                 for node in area.nodes:
                     node_item = QTreeWidgetItem(area_item)
                     if isinstance(node, TranslatorGateNode):
-                        translator = self._initial_state.patches.translator_gates[node.gate]
-                        node_item.setText(0, "{} ({})".format(node.name, translator.short_name))
+                        node_item.setText(0, "{} ({})".format(node.name, node.gate))
                     else:
                         node_item.setText(0, node.name)
                     node_item.node = node
                     if node.is_resource_node:
                         node_item.setFlags(node_item.flags() & ~Qt.ItemIsUserCheckable)
                     self._node_to_item[node] = node_item
+
+    def setup_elevators(self):
+        world_list = self.game_description.world_list
+        nodes_by_world: Dict[str, List[TeleporterNode]] = collections.defaultdict(list)
+        self._elevator_id_to_combo = {}
+
+        areas_to_not_change = {
+            2278776548,  # Sky Temple Gateway
+            2068511343,  # Sky Temple Energy Controller
+            3136899603,  # Aerie Transport Station
+            1564082177,  # Aerie
+        }
+        targets = {}
+        for world, area, node in world_list.all_worlds_areas_nodes:
+            if isinstance(node, TeleporterNode) and node.editable and area.area_asset_id not in areas_to_not_change:
+                name = world.correct_name(area.in_dark_aether)
+                nodes_by_world[name].append(node)
+                targets[f"{name} - {area.name}"] = AreaLocation(world.world_asset_id, area.area_asset_id)
+
+        if self.layout_configuration.elevators == LayoutElevators.ONE_WAY_ANYTHING:
+            targets = {}
+            for world in world_list.worlds:
+                for area in world.areas:
+                    name = world.correct_name(area.in_dark_aether)
+                    targets[f"{name} - {area.name}"] = AreaLocation(world.world_asset_id, area.area_asset_id)
+
+        combo_targets = sorted(targets.items(), key=lambda it: it[0])
+
+        for world_name in sorted(nodes_by_world.keys()):
+            nodes = sorted(nodes_by_world[world_name], key=lambda it: world_list.nodes_to_area(it).name)
+
+            group = QtWidgets.QGroupBox(self.elevators_scroll_contents)
+            group.setTitle(world_name)
+            self.elevators_scroll_layout.addWidget(group)
+            layout = QtWidgets.QGridLayout(group)
+
+            for i, node in enumerate(nodes):
+                area = world_list.nodes_to_area(node)
+                node_name = QtWidgets.QLabel(group)
+                node_name.setText(area.name)
+                layout.addWidget(node_name, i, 0)
+
+                combo = QtWidgets.QComboBox(group)
+                if self.layout_configuration.elevators == LayoutElevators.VANILLA:
+                    combo.addItem("Vanilla", node.default_connection)
+                    combo.setEnabled(False)
+                else:
+                    combo.addItem("Undefined", AreaLocation(world_list.nodes_to_world(node).world_asset_id,
+                                                            area.area_asset_id))
+                    for name, connection in combo_targets:
+                        combo.addItem(name, connection)
+
+                combo.currentIndexChanged.connect(self.update_locations_tree_for_reachable_nodes)
+                self._elevator_id_to_combo[node.teleporter_instance_id] = combo
+                layout.addWidget(combo, i, 1)
+
+    def setup_translator_gates(self):
+        world_list = self.game_description.world_list
+        resource_db = self.game_description.resource_database
+        self._translator_gate_to_combo = {}
+
+        gates = {
+            f"{area.name} ({node.gate.index})": node.gate
+            for world, area, node in world_list.all_worlds_areas_nodes
+            if isinstance(node, TranslatorGateNode)
+        }
+        translator_requirement = self.layout_configuration.translator_configuration.translator_requirement
+
+        for i, (gate_name, gate) in enumerate(sorted(gates.items(), key=lambda it: it[0])):
+            node_name = QtWidgets.QLabel(self.translator_gate_scroll_contents)
+            node_name.setText(gate_name)
+            self.translator_gate_scroll_layout.addWidget(node_name, i, 0)
+
+            combo = QtWidgets.QComboBox(self.translator_gate_scroll_contents)
+            gate_requirement = translator_requirement[gate]
+
+            if gate_requirement in (LayoutTranslatorRequirement.RANDOM,
+                                    LayoutTranslatorRequirement.RANDOM_WITH_REMOVED):
+                combo.addItem("Undefined", self._undefined_item)
+                for translator, index in translator_configuration.ITEM_INDICES.items():
+                    combo.addItem(translator.long_name, resource_db.get_item(index))
+            else:
+                combo.addItem(gate_requirement.long_name, resource_db.get_item(gate_requirement.item_index))
+                combo.setEnabled(False)
+
+            combo.currentIndexChanged.connect(self.update_locations_tree_for_reachable_nodes)
+            self._translator_gate_to_combo[gate] = combo
+            self.translator_gate_scroll_layout.addWidget(combo, i, 1)
+
+    def setup_starting_location(self, area_location: Optional[AreaLocation]):
+        world_list = self.game_description.world_list
+
+        if len(self.layout_configuration.starting_location.locations) > 1:
+            if area_location is None:
+                area_locations = sorted(self.layout_configuration.starting_location.locations,
+                                        key=lambda it: world_list.area_name(world_list.area_by_area_location(it)))
+
+                location_names = [world_list.area_name(world_list.area_by_area_location(it))
+                                  for it in area_locations]
+                selected_name = QtWidgets.QInputDialog.getItem(self, "Starting Location", "Select starting location",
+                                                               location_names, 0, False)
+                area_location = area_locations[location_names.index(selected_name[0])]
+
+            self._initial_state.node = world_list.resolve_teleporter_connection(area_location)
+
+        self._starting_nodes = {
+            node
+            for node in world_list.all_nodes
+            if node.is_resource_node and node.resource() in self._initial_state.resources
+        }
 
     def _change_item_quantity(self, pickup: PickupEntry, use_quantity_as_bool: bool, quantity: int):
         if use_quantity_as_bool:
@@ -398,7 +566,15 @@ class TrackerWindow(QMainWindow, Ui_TrackerWindow):
 
     def state_for_current_configuration(self) -> Optional[State]:
         state = self._initial_state.copy()
-        state.node = self._actions[-1]
+        if self._actions:
+            state.node = self._actions[-1]
+
+        for teleporter, combo in self._elevator_id_to_combo.items():
+            assert combo.currentData() is not None
+            state.patches.elevator_connection[teleporter] = combo.currentData()
+
+        for gate, item in self._translator_gate_to_combo.items():
+            state.patches.translator_gates[gate] = item.currentData()
 
         for pickup, quantity in self._collected_pickups.items():
             for _ in range(quantity):