+++ CHANGELOG.md
@@ -8,9 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+-   Added: The map tracker now supports random elevators, translator gates and starting location.
+
+-   Changed: The pickup spoiler in game details is now sorted.
+
 -   Fixed: Multiworld sessions should no longer occasionally duplicate messages.
 
-    Added: Starting locations tab has checkboxes to easily select all locations in an area
+-   Added: Starting locations tab has checkboxes to easily select all locations in an area
+    
+-   Fixed: Custom safe zone healing should now work in multiworld sessions.
+
 
 ## [2.0.0] - 2020-11-15